refactor: factorize utils
