+++ udata_hydra/utils/queue.py
@@ -9,7 +9,7 @@ def enqueue(fn, *args, **kwargs):
     Enqueue a task
     Simple helper useful to facilitate mock in testing
     Should be used like this (import matters):
-        from udata_hydra.utils import queue
+        from udata_hydra import queue
         queue.enqueue(fn, a, b=b, _priority="low")
     """
     priority = kwargs.pop("_priority", "default")
